Update extension.js
Added _fetchSupportedProfiles function:

This function fetches the supported profiles using supergfxctl -s.
It parses the output to get a list of supported profiles.
It then calls _addProfileToggles with the list of supported profiles and fetches the current profile.
Modified _addProfileToggles:

This function now takes a list of supported profiles and only adds those profiles to the menu.
Updated _fetchCurrentProfile and _init:

_fetchCurrentProfile is called after fetching the supported profiles.
_init initializes the profile toggles by calling _fetchSupportedProfiles.
This should dynamically update the available GPU modes based on the supported modes reported by supergfxctl -s.